# docs/api/openapi-internal.yaml
openapi: 3.1.0
info:
  title: BetTrack Dashboard REST API
  description: |
    Internal REST API for the BetTrack sports betting dashboard backend.

    ## Authentication

    The API supports two modes controlled by the `AUTH_MODE` environment variable:

    - **`none`** (default) — All routes accessible without login. Admin routes still require a valid admin session.
    - **`oauth2`** — Session-based OAuth2 authentication via Google or Microsoft (Azure AD).

    MCP integration routes and AI bet creation use API key Bearer tokens prefixed with `sk_`.

    ## Base URL

    Default: `http://localhost:3001`

    All API routes are prefixed with `/api`.
  version: 0.2.11
  contact:
    url: https://github.com/WFord26/BetTrack
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://{domain}
    description: Production
    variables:
      domain:
        default: yourdomain.com
        description: Your production domain
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: bettrack_session
      description: Session cookie established after OAuth2 login. Not required when AUTH_MODE=none.
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: sk_<key>
      description: API key with `sk_` prefix. Generate via POST /api/keys.
  schemas:
...
    BetType:
      type: string
      enum: [single, parlay, teaser]
      description: Type of bet. A single has one leg; a parlay combines multiple legs; a teaser adjusts spread/total lines by a fixed number of points.
...
    CLVReport:
      type: object
      properties:
        summary:
          type: object
          properties:
            totalBets:
              type: integer
            betsWithCLV:
              type: integer
            averageCLV:
              type: number
            positiveCLVBets:
              type: integer
            negativeCLVBets:
              type: integer
        topBets:
          type: array
          items:
            type: object
        worstBets:
          type: array
          items:
            type: object
        bySport:
          type: object
          additionalProperties:
            type: object
        byBookmaker:
          type: object
          additionalProperties:
            type: object
paths:
...
  /api/auth/google:
    get:
      tags: [Auth]
      summary: Initiate Google OAuth2 flow
      description: Redirects to Google authorization page. Requires `AUTH_MODE=oauth2` and Google credentials configured.
      parameters:
        - name: redirectTo
          in: query
          description: Path to redirect to after successful login
          schema:
            type: string
            default: /
      responses:
        '302':
          description: Redirect to Google authorization URL
        '404':
          description: Google auth not configured
...
  /api/auth/microsoft:
    get:
      tags: [Auth]
      summary: Initiate Microsoft OAuth2 flow
      description: Redirects to Microsoft authorization page. Requires `AUTH_MODE=oauth2` and Azure AD credentials configured.
      parameters:
        - name: redirectTo
          in: query
          schema:
            type: string
            default: /
      responses:
        '302':
          description: Redirect to Microsoft authorization URL
        '404':
          description: Microsoft auth not configured
...
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Log out current user
      description: Destroys the current session. Returns 200 even if auth is disabled.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    nullable: true
                  authEnabled:
                    type: boolean
        '401':
          description: Not authenticated

  # ============================================================
  # GAMES
  # ============================================================
...
  /api/analytics/clv/summary:
    get:
      tags: [CLV Analytics]
      summary: Get overall CLV summary statistics
      security:
        - sessionAuth: []
      responses:
        '200':
          description: CLV summary
  /api/analytics/clv/by-sport:
    get:
      tags: [CLV Analytics]
      summary: Get CLV breakdown by sport
      security:
        - sessionAuth: []
      responses:
        '200':
          description: CLV grouped by sport key
  /api/analytics/clv/by-bookmaker:
    get:
      tags: [CLV Analytics]
      summary: Get CLV breakdown by bookmaker
      security:
        - sessionAuth: []
      responses:
        '200':
          description: CLV grouped by bookmaker
  /api/analytics/clv/trends:
    get:
      tags: [CLV Analytics]
      summary: Get CLV trends with optional date and sport filters
      security:
        - sessionAuth: []
      parameters:
        - name: sportKey
          in: query
          schema:
            type: string
        - name: betType
          in: query
          schema:
            $ref: '#/components/schemas/BetType'
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: CLV trends with summary, top/worst bets, and breakdown by sport and bookmaker
  /api/analytics/clv/report:
    get:
      tags: [CLV Analytics]
      summary: Get full CLV report
      security:
        - sessionAuth: []
      parameters:
        - name: sportKey
          in: query
          schema:
            type: string
        - name: betType
          in: query
          schema:
            $ref: '#/components/schemas/BetType'
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Full CLV report
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/CLVReport'
  /api/analytics/clv/calculate/{betId}:
    post:
      tags: [CLV Analytics]
      summary: Calculate CLV for a specific bet
      security:
        - sessionAuth: []
      parameters:
        - name: betId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: CLV calculated successfully
  /api/analytics/clv/update-stats:
    post:
      tags: [CLV Analytics]
      summary: Recalculate and update aggregated CLV stats for current user
      security:
        - sessionAuth: []
      responses:
        '200':
          description: CLV stats updated

  # ============================================================
  # API KEYS
  # ============================================================
...